MCP/requirements.txt
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.28.0
